Add utils to compute public inputs
